Add UI logic to handle events trigger by the users
This commit includes making new variables and functions to handle events
such as clicking on the color button, eraser button, and the clear
button.
Other events such as choosing a color and canvas size, also have
corresponding functions and variables to account for these changes.
Logic is also added for the canvas section. This will allow the canvas
to change color whenever users hover and click on the canvas.